Complete authentication for Todos call

+++ server/server.js
@@ -22,10 +22,11 @@ app.use((req,res, next) => {
 
 app.use(bodyParser.json());
 
-app.post('/todos', (req, res) => {
+app.post('/todos', authenticate, (req, res) => {
 	// console.log(req.body);
 	var todo = new Todo({
-		text: req.body.text
+		text: req.body.text,
+		_creator: req.user._id
 	});
 	
 	todo.save().then( (doc) => {
@@ -35,7 +36,7 @@ app.post('/todos', (req, res) => {
 	});
 });
 
-app.patch('/todos/:id', (req, res) => {
+app.patch('/todos/:id', authenticate, (req, res) => {
 	// console.log(req.body);
 	if (!ObjectId.isValid(req.params.id)) {
 		return res.status(404).send('Invalid object ID');
@@ -56,11 +57,28 @@ app.patch('/todos/:id', (req, res) => {
 		body.completedAt = null;
 	}
 	
-	Todo.findByIdAndUpdate(req.params.id, body, { new: true }).then( (todo) => {
-		// console.log('Todo: ', todo);
+	// Todo.findByIdAndUpdate(req.params.id, body, { new: true }).then( (todo) => {
+		// if (todo) {
+			// res.status(200).send({todo});
+		// } else {
+			// res.status(204).send({todo: null});
+		// }
+	// }, (e) => {
+		// res.status(400).send(e);
+	// });
+	Todo.findById(req.params.id).then( (todo) => {
 		if (todo) {
-			res.status(200).send({todo});
-		} else {
+			if (todo._creator.equals(req.user._id)) { // check creator, before update
+				Todo.findByIdAndUpdate(req.params.id, body, { new: true }).then( (todo) => {
+					if (todo) // if item was deleted between find and find-and-delete
+						res.status(200).send({todo});
+					else
+						res.status(204).send({todo: null});
+				});
+			} else { // not the right creator; not authorized
+				res.status(401).send('Unauthorized for this Todo item');
+			} 
+		} else { // item not found
 			res.status(204).send({todo: null});
 		}
 	}, (e) => {
@@ -69,8 +87,8 @@ app.patch('/todos/:id', (req, res) => {
 });
 
 // get todos
-app.get('/todos', (req, res) => {
-	Todo.find().then( (todos) => {
+app.get('/todos', authenticate, (req, res) => {
+	Todo.find({_creator: req.user._id}).then( (todos) => {
 		res.send({todos});
 	}, (e) => {
 		res.status(400).send(e);
@@ -78,16 +96,18 @@ app.get('/todos', (req, res) => {
 });
 
 // GET by ID
-app.get('/todos/:id', (req, res) => {
-	// res.send(req.params);
+app.get('/todos/:id', authenticate, (req, res) => {
 	if (!ObjectId.isValid(req.params.id)) {
 		return res.status(404).send('Invalid object ID');
 	}
 	
 	Todo.findById(req.params.id).then( (todo) => {
-		// console.log('Todo: ', todo);
 		if (todo) {
-			res.status(200).send({todo});
+			if (todo._creator.equals(req.user._id)) {
+				res.status(200).send({todo});
+			} else {
+				res.status(401).send('Unauthorized for this todo')
+			}
 		} else {
 			res.status(204).send({todo: null});
 		}
@@ -97,15 +117,25 @@ app.get('/todos/:id', (req, res) => {
 });
 
 // DELETE by ID
-app.delete('/todos/:id', (req, res) => {
+app.delete('/todos/:id', authenticate, (req, res) => {
 	// res.send(req.params);
 	if (!ObjectId.isValid(req.params.id)) {
 		return res.status(404).send(); //'Invalid object ID');
 	}
-	Todo.findByIdAndDelete(req.params.id).then( (todo) => {
+	Todo.findById(req.params.id).then( (todo) => {
 		// console.log('Todo: ', todo);
 		if (todo) {
-			res.status(200).send({todo});
+			if (todo._creator.equals(req.user._id)) {
+				// delete
+				Todo.findByIdAndDelete(req.params.id).then( (todo) => {
+					if (todo) // if item was deleted between find and find-and-delete
+						res.status(200).send({todo});
+					else
+						res.status(204).send({todo: null});
+				});
+			} else {
+				res.status(401).send('Unauthorized for this Todo item');
+			} 
 		} else {
 			res.status(204).send({todo: null});
 		}
